push calib routines into class modules
