+++ pypit/arms.py
@@ -16,7 +16,15 @@ from pypit import arsciexp
 from pypit.core import arsetup
 from pypit import arpixels
 from pypit.core import arsort
+from pypit import arcimage
+from pypit import biasframe
+from pypit import bpmimage
+from pypit import flatfield
 from pypit import fluxspec
+from pypit import traceslits
+from pypit import wavecalib
+from pypit import waveimage
+from pypit import wavetilts
 
 from pypit import ardebug as debugger
 
@@ -132,7 +140,7 @@ def ARMS(spectrograph, fitstbl, setup_dict):
             else:
                 # Grab it
                 #   Bias will either be an image (ndarray) or a command (str, e.g. 'overscan') or none
-                msbias, _ = arcalib.get_msbias(det, setup, sci_ID, fitstbl, tsettings)
+                msbias, _ = biasframe.get_msbias(det, setup, sci_ID, fitstbl, tsettings)
                 # Save
                 calib_dict[setup]['bias'] = msbias
 
@@ -142,7 +150,7 @@ def ARMS(spectrograph, fitstbl, setup_dict):
                 msarc = calib_dict[setup]['arc']
             else:
                 # Grab it -- msarc will be a 2D image
-                msarc, _ = arcalib.get_msarc(det, setup, sci_ID, spectrograph, fitstbl, tsettings, msbias)
+                msarc, _ = arcimage.get_msarc(det, setup, sci_ID, spectrograph, fitstbl, tsettings, msbias)
                 # Save
                 calib_dict[setup]['arc'] = msarc
 
@@ -152,7 +160,7 @@ def ARMS(spectrograph, fitstbl, setup_dict):
                 msbpm = calib_dict[setup]['bpm']
             else:
                 # Grab it -- msbpm is a 2D image
-                msbpm, _ = arcalib.get_mspbm(det, spectrograph, tsettings, msarc.shape,
+                msbpm, _ = bpmimage.get_mspbm(det, spectrograph, tsettings, msarc.shape,
                                       binning=fitstbl['binning'][scidx],
                                       reduce_badpix=settings.argflag['reduce']['badpix'],
                                       msbias=msbias)
@@ -172,7 +180,7 @@ def ARMS(spectrograph, fitstbl, setup_dict):
                 ts_settings = dict(trace=settings.argflag['trace'], masters=settings.argflag['reduce']['masters'])
                 ts_settings['masters']['directory'] = settings.argflag['run']['directory']['master']+'_'+ settings.argflag['run']['spectrograph']
                 # Get it -- Key arrays are in the tslits_dict
-                tslits_dict, _ = arcalib.get_tslits_dict(
+                tslits_dict, _ = traceslits.get_tslits_dict(
                     det, setup, spectrograph, sci_ID, ts_settings, tsettings, fitstbl, pixlocn,
                     msbias, msbpm, datasec_img, trim=settings.argflag['reduce']['trim'])
                 # Save in calib
@@ -198,7 +206,7 @@ def ARMS(spectrograph, fitstbl, setup_dict):
                 wvc_settings['masters']['directory'] = settings.argflag['run']['directory']['master']+'_'+ settings.argflag['run']['spectrograph']
                 nonlinear = settings.spect[settings.get_dnum(det)]['saturation'] * settings.spect[settings.get_dnum(det)]['nonlinear']
                 # Get it
-                wv_calib, wv_maskslits, _ = arcalib.get_wv_calib(
+                wv_calib, wv_maskslits, _ = wavecalib.get_wv_calib(
                     det, setup, spectrograph, sci_ID, wvc_settings, fitstbl, tslits_dict, pixlocn,
                     msarc, nonlinear=nonlinear)
                 # Save in calib
@@ -219,7 +227,7 @@ def ARMS(spectrograph, fitstbl, setup_dict):
                 tilt_settings['tilts']['function'] = settings.argflag['trace']['slits']['function']
                 tilt_settings['masters']['directory'] = settings.argflag['run']['directory']['master']+'_'+ settings.argflag['run']['spectrograph']
                 # Get it
-                mstilts, wt_maskslits, _ = arcalib.get_wv_tilts(
+                mstilts, wt_maskslits, _ = wavetilts.get_wv_tilts(
                     det, setup, tilt_settings, settings_det, tslits_dict, pixlocn,
                     msarc, wv_calib, maskslits)
                 # Save
@@ -244,7 +252,7 @@ def ARMS(spectrograph, fitstbl, setup_dict):
                                          detector=settings.spect[dnum])
                     flat_settings['masters']['directory'] = settings.argflag['run']['directory']['master']+'_'+ settings.argflag['run']['spectrograph']
                     # Get it
-                    mspixflatnrm, slitprof, _ = arcalib.get_msflat(
+                    mspixflatnrm, slitprof, _ = flatfield.get_msflat(
                         det, setup, spectrograph, sci_ID, fitstbl, tslits_dict, datasec_img,
                         flat_settings, msbias, mstilts)
                     # Save internallly
@@ -267,7 +275,7 @@ def ARMS(spectrograph, fitstbl, setup_dict):
                     wvimg_settings = dict(masters=settings.argflag['reduce']['masters'].copy())
                     wvimg_settings['masters']['directory'] = settings.argflag['run']['directory']['master']+'_'+ settings.argflag['run']['spectrograph']
                     # Get it
-                    mswave, _ = arcalib.get_mswave(
+                    mswave, _ = waveimage.get_mswave(
                         setup, tslits_dict, wvimg_settings, mstilts, wv_calib, maskslits)
                 # Save internally
                 calib_dict[setup]['wave'] = mswave
